requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.1

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
httpx==0.26.0
coverage==7.4.1 